add a toggle for disabling fingerprint unlock
Ported from 11. In 12 fingerprint is handled by the new biometric API
and so technically in 12 this is not just disabling fingerprint unlock
but all biometrics. This does not affect the Pixel 4 generation of
devices with face scanning however as they were released on Q and so
use the legacy face API.
No frameworks/base changes are needed for 12 in this case other than
migration which can be done in another commit.

Change-Id: I7aa082c5e0bc747d40177755c4cefbc6eb47e482